+++ views/front/frontNounouProfilView.php
@@ -18,7 +18,7 @@
                         <label for="pseudo">Pseudo  : </label> 
                     </td>
                     <td>
-                        <input class="form-control" type="text" name="pseudo" id="pseudo" value="<?= $nounou->pseudo(); ?>" required> 
+                        <input class="form-control" type="text" name="pseudo" id="pseudo" value="<?= htmlspecialchars($nounou->pseudo()); ?>" required> 
                     </td>
                 </tr>
                 <tr>
@@ -26,7 +26,7 @@
                         <label for="nom">Nom  : </label> 
                     </td>
                     <td>
-                        <input class="form-control" type="text" name="nom" id="nom" value="<?= $nounou->nom(); ?>" required> 
+                        <input class="form-control" type="text" name="nom" id="nom" value="<?= htmlspecialchars($nounou->nom()); ?>" required> 
                     </td>
                 </tr>
                 <tr>
@@ -34,7 +34,7 @@
                         <label for="prenom">Prenom  : </label> 
                     </td>
                     <td>
-                        <input class="form-control" type="text" name="prenom" id="prenom" value="<?= $nounou->prenom(); ?>" required>
+                        <input class="form-control" type="text" name="prenom" id="prenom" value="<?= htmlspecialchars($nounou->prenom()); ?>" required>
                     </td>
                 </tr>
                 <tr>
@@ -42,7 +42,7 @@
                         <label for="email">Adresse email  : </label> 
                     </td>
                     <td>
-                        <input class="form-control" type="email" name="email" id="email" value="<?= $nounou->email(); ?>" required>
+                        <input class="form-control" type="email" name="email" id="email" value="<?= htmlspecialchars($nounou->email()); ?>" required>
                     </td>
                 </tr>
                 <tr>
@@ -50,7 +50,7 @@
                        <label for="confirm_email">Confirmation adresse email  : </label> 
                     </td>
                     <td>
-                        <input  class="form-control" type="email" name="confirm email" id="confirm email" value="<?= $nounou->email(); ?>" required>
+                        <input  class="form-control" type="email" name="confirm email" id="confirm email" value="<?= htmlspecialchars($nounou->email()); ?>" required>
                     </td>
                 </tr>
                 <tr>
@@ -58,7 +58,7 @@
                        <label for="experience">Année(s) d'expérience  : </label> 
                     </td>
                     <td>
-                        <input class="form-control" type="number" name="experience" id="experience" value="<?= $nounou->experience(); ?>" required>
+                        <input class="form-control" type="number" name="experience" id="experience" value="<?= htmlspecialchars($nounou->experience()); ?>" required>
                     </td>
                 </tr>
 
@@ -67,7 +67,7 @@
                        <label for="dispo">Place(s) disponible(s)  : </label> 
                     </td>
                     <td>
-                        <input class="form-control" type="number" name="place_dispo" id="dispo" value="<?= $nounou->place_dispo(); ?>" required>
+                        <input class="form-control" type="number" name="place_dispo" id="dispo" value="<?= htmlspecialchars($nounou->place_dispo()); ?>" required>
                     </td>
                 </tr>
 
@@ -94,19 +94,14 @@
                     </td>
                     <td>
                         <select class="cityContainer" type="text" name="ville" id="ville" required> 
-                            <option><?= $nounou->ville(); ?></option>
+                            <option><?= htmlspecialchars($nounou->ville()); ?></option>
                         </select>
                     </td>
                 </tr>
             </table>
         <input class="btn btn-primary" type="submit" value="Enregistrer"/>
     </form>
-
-
-
-    <?php if(isset($_SESSION['editNounou_message'])) {
-        echo $_SESSION['editNounou_message'];
-    } ?>
+    <?php if(isset($_SESSION['editNounou_message'])) {echo $_SESSION['editNounou_message'];} ?>
 </div>
 
 <hr>
@@ -114,17 +109,13 @@
 <div class="formulaire_profil" id="updatePassword">
     <p class="lead">Changer de mot de passe</p>
     <?php if(isset($_SESSION['editPasswordNounou'])) {echo $_SESSION['editPasswordNounou'];} ?>
-
-
-    <form action="index.php?action=updatePasswordNounou&amp;pseudo=<?= $nounou->pseudo(); ?>" method="POST">
+    <form action="index.php?action=updatePasswordNounou&amp;pseudo=<?= htmlspecialchars($nounou->pseudo()); ?>" method="POST">
         <label for="password"> Définir mon mot de passe : <input class="form-control" type="password" name="password" id="password" required> </label>  <br/> 
         <label for="confirm_password"> Confirmation mot de passe  : <input class="form-control" type="password" name="confirm password" id="confirm password" required> </label> <br/> 
-        <input class="btn btn-primary"  type="submit" value="Modifier le mot de passe">  <br>
-        <a id="deleteProfil" class="btn btn-danger" href="index.php?action=deleteNounou&amp;idNounou=<?= $nounou->id(); ?>">Supprimer mon profil</a>
+        <input class="btn btn-primary"  type="submit" value="Modifier le mot de passe" onclick="return confirm('Etes-vous sûr de vouloir modifier votre mot de passe ?');">  <br>
+        <a id="deleteProfil" class="btn btn-danger" href="index.php?action=deleteNounou&amp;idNounou=<?= htmlspecialchars($nounou->id()); ?>" onclick="return confirm('Etes-vous sûr de vouloir supprimer votre profil ?');">Supprimer mon profil</a>
     </form>
 </div>
-
-
 
 <?php $content = ob_get_clean(); ?>
